반차까지

+++ src/main/java/com/mycompany/webapp/vacation/model/VacationList.java
@@ -9,7 +9,7 @@ public class VacationList {
 	private int vacationId;
 	private String name;
 	private String vacationName;
-	private int countDay;
+	private double countDay;
 	private String startDate;
 	private String endDate;
 	private String status;
